requirements.txt
openai 
python-docx 
sentence-transformers 
numpy
pdfplumber